Moved connection.py to volttron/platform/vip/agent and updated links.
