+++ services/core/VolttronCentralPlatform/vcplatform/agent.py
@@ -59,42 +59,36 @@
 from __future__ import absolute_import, print_function
 
 import base64
-import requests
-import tempfile
-from copy import deepcopy
 import datetime
 import hashlib
 import logging
-import os
 import re
 import shutil
 import sys
-import uuid
+import tempfile
 import urlparse
+from copy import deepcopy
 
 import gevent
 import gevent.event
 import psutil
-
+import requests
 from volttron.platform import get_home
-from volttron.platform.agent.utils import (
-    get_aware_utc_now, format_timestamp, parse_timestamp_string)
-from volttron.platform.messaging.topics import (LOGGER, PLATFORM_VCP_DEVICES,
-                                                PLATFORM)
-from volttron.platform.vip.agent.subsystems.query import Query
-from volttron.platform.vip.connection import Connection
-
-from volttron.platform.vip.agent import *
-
 from volttron.platform import jsonrpc
-from volttron.platform.auth import AuthEntry, AuthFile
 from volttron.platform.agent import utils
 from volttron.platform.agent.known_identities import (
     VOLTTRON_CENTRAL, VOLTTRON_CENTRAL_PLATFORM)
-from volttron.platform.messaging.health import UNKNOWN_STATUS, Status, \
+from volttron.platform.agent.utils import (
+    get_aware_utc_now, format_timestamp, parse_timestamp_string)
+from volttron.platform.auth import AuthEntry, AuthFile
+from volttron.platform.jsonrpc import (INTERNAL_ERROR, INVALID_PARAMS)
+from volttron.platform.messaging.health import Status, \
     GOOD_STATUS, BAD_STATUS
-from volttron.platform.jsonrpc import (INTERNAL_ERROR, INVALID_PARAMS,
-                                       METHOD_NOT_FOUND)
+from volttron.platform.messaging.topics import (LOGGER, PLATFORM_VCP_DEVICES,
+                                                PLATFORM)
+from volttron.platform.vip.agent import *
+from volttron.platform.vip.agent.connection import Connection
+from volttron.platform.vip.agent.subsystems.query import Query
 from volttron.utils.persistance import load_create_store
 
 __version__ = '3.5.4'
